requirements.txt
fastapi 
uvicorn 
torch 
torchvision 
pillow 
jinja2 
python-multipart